Moved member service with the other services. Preparing for library functionality.
