+++ src/app/_services/member.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from 'src/environments/environment';
-import { Member } from './_models/member';
+import { Member } from '../_models/member';
 
 @Injectable({
   providedIn: 'root'
