+++ src/app/admin/_modals/library-access-modal/library-access-modal.component.html
@@ -1,1 +1,22 @@
-<p>library-access-modal works!</p>
+
+<div class="modal-header">
+    <h4 class="modal-title" id="modal-basic-title">Library Access</h4>
+    <button type="button" class="close" aria-label="Close" (click)="close()">
+    <span aria-hidden="true">&times;</span>
+    </button>
+</div>
+<div class="modal-body">
+    <div class="list-group">
+        <li class="list-group-item">
+            <div class="form-check">
+                <input id="library1" type="checkbox" aria-label="Library 1" class="form-check-input">
+                <label for="library1" class="form-check-label">Library 1</label>
+            </div>
+        </li>
+    </div>
+</div>
+<div class="modal-footer">
+    <button type="button" class="btn btn-light" (click)="reset()">Cancel</button>
+    <button type="button" class="btn btn-secondary" (click)="close()">Cancel</button>
+    <button type="button" class="btn btn-primary" (click)="save()">Save</button>
+</div>
