+++ src/app/library/library.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { take } from 'rxjs/operators';
-import { MemberService } from '../member.service';
 import { Library } from '../_models/library';
 import { User } from '../_models/user';
 import { AccountService } from '../_services/account.service';
 import { LibraryService } from '../_services/library.service';
+import { MemberService } from '../_services/member.service';
 
 @Component({
   selector: 'app-library',
